requirements.txt
# =============================
# Core Python & Virtual Env
# =============================
pip>=24.0
setuptools>=70.0
wheel>=0.43.0
virtualenv>=20.25.0

# =============================
# Core UI
# =============================
gradio==3.50.2

# =============================
# Image Processing
# =============================
Pillow>=10.4.0,<11.0.0
opencv-python>=4.10.0,<5.0.0
scikit-image>=0.24.0,<1.0.0

# =============================
# Numerical / Data
# =============================
# Gradio 3.50.2 is incompatible with NumPy 2.x
numpy==1.26.4
matplotlib>=3.9.2,<4.0.0
tqdm>=4.66.5,<5.0.0
joblib>=1.4.2,<2.0.0
pandas>=2.2.2,<3.0.0

# =============================
# Deep Learning / GPU
# =============================
# Installed per-OS in the workflow; these ranges keep pip happy.
# 
# For NVIDIA CUDA GPUs:
#   pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu121
#
# For AMD ROCm GPUs (Linux):
#   pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/rocm6.0
#
# For macOS (MPS/Metal backend):
#   pip install torch torchvision torchaudio
torch>=2.3.1,<2.5.0
torchvision>=0.18.0,<0.20.0
torchaudio>=2.3.1,<2.5.0

# =============================
# Backend / API / Networking
# =============================
requests>=2.32.3,<3.0.0
httpx>=0.27.0,<0.28.0
psutil>=6.0.0,<7.0.0

# =============================
# Extra Utilities
# =============================
ujson>=5.10.0,<6.0.0
orjson>=3.10.7,<4.0.0
platformdirs>=4.3.6,<5.0.0

# =============================
# Optional but useful
# =============================
rich>=13.9.2,<14.0.0
loguru>=0.7.2,<1.0.0
